Move GeckoLibNetwork to GeckoLibUtil
