+++ neoforge/src/main/java/software/bernie/geckolib/network/packet/AnimDataSyncPacket.java
@@ -11,6 +11,7 @@ import software.bernie.geckolib.core.animatable.GeoAnimatable;
 import software.bernie.geckolib.network.GeckoLibNetwork;
 import software.bernie.geckolib.network.SerializableDataTicket;
 import software.bernie.geckolib.util.ClientUtils;
+import software.bernie.geckolib.util.GeckoLibUtil;
 
 /**
  * Packet for syncing user-definable animation data for {@link SingletonGeoAnimatable} instances
@@ -42,7 +43,7 @@ public record AnimDataSyncPacket<D>(String syncableId, long instanceId, Serializ
 
 	public void receivePacket(PlayPayloadContext context) {
 		context.workHandler().execute(() -> {
-			GeoAnimatable animatable = GeckoLibNetwork.getSyncedAnimatable(this.syncableId);
+			GeoAnimatable animatable = GeckoLibUtil.getSyncedAnimatable(this.syncableId);
 
 			if (animatable instanceof SingletonGeoAnimatable singleton)
 				singleton.setAnimData(ClientUtils.getClientPlayer(), this.instanceId, this.dataTicket, this.data);
